Retire kokkos llvm build on ORNL Jenkins (#7801)
* Use Clang 18 in CUDA+Clang build on ORNL Jenkins

Retire Kokkos' LLVM build from our fork.  Use Clang installed from the
system package manager instead.
I chose Clang 18 since version 12 warns that the latest supported
version of CUDA is 10.1
Our fork of LLVM is wildly outdated.

* Use DEBIAN_FRONTEND=noninteractive when installing packages

* Truncate Clang version since we do not have control over it


* Truncate Clang version since we have no control over it


* Fix pipeline (use additionalBuildArgs instead of args)

* Fix modernize-type-traits Clang-Tidy check with CUDA

* Temporarily diable bugprone-unchecked-optional-access check in CUDA Graph interop tests

* fixup! Fix modernize-type-traits Clang-Tidy check with CUDA

* Fix bugprone-implicit-widening-of-multiplication-result for CUDA

* Fix clang-diagnostic-deprecated-copy warning with CUDA

* [DO NOT MERGE] try Debug build to get backtraces

* fixup! Fix modernize-type-traits Clang-Tidy check with CUDA

* Change to CUDA 11.8 and Clang 15

* Don't use debug build because that makes one of the test fail

There is a PTX error during assembly.

* Disable -Wpass-failed=... warnings

---------

+++ core/src/Cuda/Kokkos_Cuda_Instance.cpp
@@ -304,11 +304,10 @@ void CudaInternal::initialize(cudaStream_t stream) {
   {
     // Maximum number of warps,
     // at most one warp per thread in a warp for reduction.
-    auto const maxWarpCount = std::min<unsigned>(
-        m_deviceProp.maxThreadsPerBlock / CudaTraits::WarpSize,
-        CudaTraits::WarpSize);
-    unsigned const reduce_block_count =
-        maxWarpCount * Impl::CudaTraits::WarpSize;
+    auto const maxWarpCount =
+        std::min<size_t>(m_deviceProp.maxThreadsPerBlock / CudaTraits::WarpSize,
+                         CudaTraits::WarpSize);
+    size_t const reduce_block_count = maxWarpCount * Impl::CudaTraits::WarpSize;
 
     (void)scratch_unified(16 * sizeof(size_type));
     (void)scratch_flags(reduce_block_count * 2 * sizeof(size_type));
